Move the hellfire hud to the laser module
