+++ addons/hellfire/XEH_PREP.hpp
@@ -2,4 +2,3 @@ LOG("prep");
 PREP(attackProfile);
 PREP(getAttackProfileSettings);
 PREP(setupVehicle);
-PREP(showHud);
