+++ addons/hellfire/config.cpp
@@ -20,4 +20,3 @@ class CfgPatches {
 #include "CfgMagazines.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
-#include "RscTitles.hpp"
