+++ addons/hellfire/script_component.hpp
@@ -20,8 +20,3 @@
 #define STAGE_SEEK_CRUISE 2
 #define STAGE_ATTACK_CRUISE 3
 #define STAGE_ATTACK_TERMINAL 4
-
-#define IDC_MODECONTROLGROUP 1000
-#define IDC_ATTACKMODE 1001
-#define IDC_LASERCODE 1002
-#define IDC_LASERICON 1003
