+++ addons/laser/config.cpp
@@ -13,8 +13,19 @@ class CfgPatches {
     };
 };
 
-#include "ACE_Settings.hpp"
+#include "ACE_Settings.hpp" 
 #include "CfgEventhandlers.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
+
+
+class RscControlsGroup;
+class VScrollbar;
+class HScrollbar;
+class RscText;
+class RscMapControl;
+class RscControlsGroupNoScrollbars;
+class RscPictureKeepAspect;
+
 #include "RscInGameUI.hpp"
+#include "RscTitles.hpp"
